+++ src/hooks/data/useCollectionFetch.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { Entity, EntitySchema, FilterValues } from "../../models";
 import { useDataSource } from "./useDataSource";
 
@@ -74,6 +74,9 @@ export function useCollectionFetch<M extends { [Key: string]: any }>(
     const initialEntities = entitiesDisplayedFirst ? entitiesDisplayedFirst.filter(e => !!e.values) : [];
     const [data, setData] = useState<Entity<M>[]>(initialEntities);
 
+    // this is a hack to prevent false updates from Firestore, which returns less data than requested for some reason
+    const resetData = useRef<boolean>(false);
+
     const [dataLoading, setDataLoading] = useState<boolean>(false);
     const [dataLoadingError, setDataLoadingError] = useState<Error | undefined>();
     const [noMoreToLoad, setNoMoreToLoad] = useState<boolean>(false);
@@ -89,15 +92,23 @@ export function useCollectionFetch<M extends { [Key: string]: any }>(
 
 
     useEffect(() => {
+        resetData.current = true;
+    }, [path, itemCount, currentSort, sortByProperty, filterValues, searchString]);
+
+    useEffect(() => {
 
         setDataLoading(true);
 
         const onEntitiesUpdate = (entities: Entity<M>[]) => {
             setDataLoading(false);
             setDataLoadingError(undefined);
-            updateData(entities);
+            if (resetData.current || entities.length >= data.length) {
+                resetData.current = false;
+                updateData(entities);
+            }
             setNoMoreToLoad(!itemCount || entities.length < itemCount);
         };
+
         const onError = (error: Error) => {
             console.error("ERROR", error);
             setDataLoading(false);